requirements.txt
Django==6.0.3
djangorestframework==3.15.2
mysqlclient==2.2.4
reportlab==4.2.2
